+++ 2-3-PipelineCache/code/main.cpp
@@ -334,7 +334,7 @@ int main(int argc, char* argv[])
 					pipeline
 				);
 			chrono::time_point compileEnd = chrono::high_resolution_clock::now();
-			double delta = chrono::duration<double>(compileEnd - compileStart).count();
+			float delta = chrono::duration<float>(compileEnd - compileStart).count();
 			if(r == vk::Result::eSuccess)
 				cout << " done.\n   The pipeline was retrieved from a cache in " << delta * 1e3 << "ms." << endl;
 			else if(r == vk::Result::ePipelineCompileRequired)
@@ -364,7 +364,7 @@ int main(int argc, char* argv[])
 					}
 				);
 			chrono::time_point compileEnd = chrono::high_resolution_clock::now();
-			double delta = chrono::duration<double>(compileEnd - compileStart).count();
+			float delta = chrono::duration<float>(compileEnd - compileStart).count();
 			if(pipelineCacheControlSupport)
 				// pipeline was compiled - we know it from pipeline cache control
 				cout << " done.\n   The pipeline was compiled in " << delta * 1e3 << "ms." << endl;
@@ -461,10 +461,10 @@ int main(int argc, char* argv[])
 		cout << "Done." << endl;
 
 		// print results
-		double delta = chrono::duration<double>(t2 - t1).count();
+		float delta = chrono::duration<float>(t2 - t1).count();
 		cout << "Computation time: " << delta * 1e3 << "ms." << endl;
 		constexpr uint64_t numInstructions = uint64_t(20000) * 128 * workgroupCountX * workgroupCountY * workgroupCountZ;
-		cout << "Computing performance: " << double(numInstructions) / delta * 1e-12 << " TFLOPS." << endl;
+		cout << "Computing performance: " << float(numInstructions) / delta * 1e-12 << " TFLOPS." << endl;
 
 	// catch exceptions
 	} catch(vk::Error& e) {
